requirements.txt
aniso8601==8.0.0
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-RESTful==0.3.8
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.7.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
protobuf==3.11.3
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
Werkzeug==1.0.1
